+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/TestSuiteBase.java
@@ -32,6 +32,7 @@ import com.azure.cosmos.implementation.HttpConstants;
 import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
 import com.azure.cosmos.implementation.InternalObjectNode;
 import com.azure.cosmos.implementation.PathParser;
+import com.azure.cosmos.implementation.QueryFeedOperationState;
 import com.azure.cosmos.implementation.Resource;
 import com.azure.cosmos.implementation.TestConfigurations;
 import com.azure.cosmos.implementation.Utils;
@@ -95,8 +96,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.spy;
 
-@Listeners({TestNGLogListener.class})
-public class TestSuiteBase extends CosmosAsyncClientTest {
+public abstract class TestSuiteBase extends CosmosAsyncClientTest {
 
     private static final int DEFAULT_BULK_INSERT_CONCURRENCY_LEVEL = 500;
     private static final ObjectMapper objectMapper = new ObjectMapper();
@@ -205,7 +205,7 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
 
     @BeforeSuite(groups = {"thinclient", "fast", "long", "direct", "multi-region", "multi-master", "flaky-multi-master", "emulator",
         "emulator-vnext", "split", "query", "cfp-split", "circuit-breaker-misc-gateway", "circuit-breaker-misc-direct",
-        "circuit-breaker-read-all-read-many", "fi-multi-master", "long-emulator", "fi-thinclient-multi-region", "fi-thinclient-multi-master"}, timeOut = SUITE_SETUP_TIMEOUT)
+        "circuit-breaker-read-all-read-many", "fi-multi-master", "long-emulator", "fi-thinclient-multi-region", "fi-thinclient-multi-master", "fault-injection-barrier"}, timeOut = SUITE_SETUP_TIMEOUT)
     public void beforeSuite() {
 
         logger.info("beforeSuite Started");
@@ -221,16 +221,9 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
         }
     }
 
-    @BeforeSuite(groups = {"unit"})
-    public void parallelizeUnitTests(ITestContext context) {
-        // TODO: Parallelization was disabled due to flaky tests. Re-enable after fixing the flaky tests.
-//        context.getSuite().getXmlSuite().setParallel(XmlSuite.ParallelMode.CLASSES);
-//        context.getSuite().getXmlSuite().setThreadCount(Runtime.getRuntime().availableProcessors());
-    }
-
     @AfterSuite(groups = {"thinclient", "fast", "long", "direct", "multi-region", "multi-master", "flaky-multi-master",
         "emulator", "split", "query", "cfp-split", "circuit-breaker-misc-gateway", "circuit-breaker-misc-direct",
-        "circuit-breaker-read-all-read-many", "fi-multi-master", "long-emulator", "fi-thinclient-multi-region", "fi-thinclient-multi-master"}, timeOut = SUITE_SHUTDOWN_TIMEOUT)
+        "circuit-breaker-read-all-read-many", "fi-multi-master", "long-emulator", "fi-thinclient-multi-region", "fi-thinclient-multi-master", "fault-injection-barrier"}, timeOut = SUITE_SHUTDOWN_TIMEOUT)
     public void afterSuite() {
 
         logger.info("afterSuite Started");
@@ -287,28 +280,51 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
     }
 
     protected static void truncateCollection(CosmosAsyncContainer cosmosContainer) {
-        int i = 0;
-        while (i < 100) {
-            try {
-                truncateCollectionInternal(cosmosContainer);
-                return;
-            } catch (CosmosException exception) {
-                if (exception.getStatusCode() != HttpConstants.StatusCodes.TOO_MANY_REQUESTS
-                    || exception.getSubStatusCode() != 3200) {
 
-                    logger.error("No retry of exception", exception);
-                    throw exception;
-                }
-
-                i++;
-                logger.info("Retrying truncation after 100ms - iteration " + i);
+        try {
+            int i = 0;
+            while (i < 100) {
                 try {
-                    Thread.sleep(100);
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
+                    truncateCollectionInternal(cosmosContainer);
+                    return;
+                } catch (CosmosException exception) {
+                    if (exception.getStatusCode() != HttpConstants.StatusCodes.TOO_MANY_REQUESTS
+                        || exception.getSubStatusCode() != 3200) {
+
+                        logger.error("No retry of exception", exception);
+                        throw exception;
+                    }
+
+                    i++;
+                    logger.info("Retrying truncation after 100ms - iteration " + i);
+                    try {
+                        Thread.sleep(100);
+                    } catch (InterruptedException e) {
+                        throw new RuntimeException(e);
+                    }
                 }
             }
+        } finally {
+            // TODO @fabianm - Resetting leak detection - there is some flakiness when tests
+            // truncate collection and hit throttling - this will need to be investigated
+            // separately
+            CosmosNettyLeakDetectorFactory.resetIdentifiedLeaks();
         }
+    }
+
+    protected static void expectCount(CosmosAsyncContainer cosmosContainer, int expectedCount) {
+        CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
+        options.setCosmosEndToEndOperationLatencyPolicyConfig(
+            new CosmosEndToEndOperationLatencyPolicyConfigBuilder(Duration.ofHours(1))
+                .build()
+        );
+        options.setMaxDegreeOfParallelism(-1);
+        List<Integer> counts = cosmosContainer
+            .queryItems("SELECT VALUE COUNT(0) FROM root", options, Integer.class)
+            .collectList()
+            .block();
+        assertThat(counts).hasSize(1);
+        assertThat(counts.get(0)).isEqualTo(expectedCount);
     }
 
     private static void truncateCollectionInternal(CosmosAsyncContainer cosmosContainer) {
@@ -349,6 +365,9 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
 
                            return cosmosContainer.deleteItem(doc.getId(), partitionKey);
                        }).then().block();
+
+        expectCount(cosmosContainer, 0);
+
         logger.info("Truncating collection {} triggers ...", cosmosContainerId);
 
         cosmosContainer.getScripts().queryTriggers("SELECT * FROM root", options)
@@ -924,15 +943,9 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
         }
     }
 
-    static protected void safeCloseAsync(CosmosAsyncClient client) {
-        if (client != null) {
-            new Thread(() -> {
-                try {
-                    client.close();
-                } catch (Exception e) {
-                    logger.error("failed to close client", e);
-                }
-            }).start();
+    static protected void safeClose(QueryFeedOperationState state) {
+        if (state != null) {
+            safeClose(state.getClient());
         }
     }
 
@@ -1580,4 +1593,28 @@ public class TestSuiteBase extends CosmosAsyncClientTest {
         }
         return outputStream.toByteArray();
     }
+
+    public static String captureNettyLeaks() {
+        System.gc();
+        try {
+            Thread.sleep(5_000);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+        List<String> nettyLeaks = CosmosNettyLeakDetectorFactory.resetIdentifiedLeaks();
+        if (nettyLeaks.size() > 0) {
+            StringBuilder sb = new StringBuilder();
+            sb.append("\n");
+            for (String leak : nettyLeaks) {
+                sb.append(leak).append("\n");
+            }
+
+            return "NETTY LEAKS detected: "
+                + "\n\n"
+                + sb;
+        }
+
+        return "";
+    }
+
 }