+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/BlobApiTests.java
@@ -3173,4 +3173,19 @@ public class BlobApiTests extends BlobTestBase {
         deleteFileShareWithoutDependency(shareName);
     }
 
+    @Test
+    public void blobNameEncodingOnGetBlobUrl() {
+        BlobClient blobClient = cc.getBlobClient("my blob");
+        String expectedEncodedBlobName = "my%20blob";
+        assertTrue(blobClient.getBlobUrl().contains(expectedEncodedBlobName));
+    }
+
+    @Test
+    void containerNameEncodingOnGetBlobUrl() {
+        BlobContainerClient containerClient = primaryBlobServiceClient.getBlobContainerClient("my container");
+        BlobClient blobClient = containerClient.getBlobClient(generateBlobName());
+        String expectedEncodedContainerName = "my%20container";
+        assertTrue(blobClient.getBlobUrl().contains(expectedEncodedContainerName));
+    }
+
 }