+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/ContainerApiTests.java
@@ -2005,6 +2005,16 @@ public class ContainerApiTests extends BlobTestBase {
         assertTrue(exception.getMessage().contains(INVALID_VERSION_HEADER_MESSAGE));
     }
 
+    // Tests that the container name is URL encoded. Container names with special characters are not supported
+    // by the service, however, the names should still be encoded.
+    @Test
+    public void getBlobContainerUrlEncodesContainerName() {
+        String containerName = "my container";
+        BlobContainerClient containerClient = primaryBlobServiceClient.getBlobContainerClient(containerName);
+
+        assertTrue(containerClient.getBlobContainerUrl().contains("my%20container"));
+    }
+
     // TODO: Reintroduce these tests once service starts supporting it.
 
     //    public void Rename() {