+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/ContainerAsyncApiTests.java
@@ -2132,4 +2132,14 @@ public class ContainerAsyncApiTests extends BlobTestBase {
         });
     }
 
+    // Tests that the container name is URL encoded. Container names with special characters are not supported
+    // by the service, however, the names should still be encoded.
+    @Test
+    public void getBlobContainerUrlEncodesContainerName() {
+        String containerName = "my container";
+        BlobContainerAsyncClient containerClient
+            = primaryBlobServiceAsyncClient.getBlobContainerAsyncClient(containerName);
+
+        assertTrue(containerClient.getBlobContainerUrl().contains("my%20container"));
+    }
 }