+++ sdk/storage/azure-storage-file-datalake/src/test/java/com/azure/storage/file/datalake/FileApiTest.java
@@ -3599,4 +3599,11 @@ public class FileApiTest extends DataLakeTestBase {
     public void pathGetSystemPropertiesFileMin() {
         assertNotNull(fc.getSystemProperties());
     }
+
+    @Test
+    public void fileNameEncodingOnGetPathUrl() {
+        DataLakeFileClient fileClient = dataLakeFileSystemClient.getFileClient("my file");
+        String expectedName = "my%20file";
+        assertTrue(fileClient.getPathUrl().contains(expectedName));
+    }
 }