+++ sdk/storage/azure-storage-blob-batch/CHANGELOG.md
@@ -7,8 +7,10 @@
 ### Breaking Changes
 
 ### Bugs Fixed
+- Fixed an issue where `BlobBatchSetBlobAccessTierOptions` would not properly handle blob names with special characters.
 
 ### Other Changes
+- Added support for container names with special characters when using OneLake.
 
 ## 12.28.0 (2025-10-21)